+++ ADOL-C/src/hos_reverse.cpp
@@ -1,8 +1,8 @@
 /*----------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
- File:     hov_reverse.c
+ File:     hos_reverse.cpp
  Revision: $Id$
- Contents: hov_reverse (higher-order-vector reverse mode)
+ Contents: hos_reverse (higher-order-scalar reverse mode)
 
  Copyright (c) Andrea Walther, Andreas Griewank, Andreas Kowarz,
                Hristo Mitev, Sebastian Schlenkrich, Jean Utke, Olaf Vogel
@@ -13,6 +13,6 @@
 
 ----------------------------------------------------------------------------*/
 
-#define _HOV_ 1
-#include <ho_rev.c>
-#undef _HOV_
+#define _HOS_ 1
+#include <ho_rev.cpp>
+#undef _HOS_
